+++ src/lib/config.ts
@@ -60,7 +60,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "latest",
-    sdkVersion: "0.6.0",
-    genVersion: "2.392.0",
-    userAgent: "speakeasy-sdk/typescript 0.6.0 2.392.0 latest @lukehagar/discoursejs",
+    sdkVersion: "0.6.1",
+    genVersion: "2.393.4",
+    userAgent: "speakeasy-sdk/typescript 0.6.1 2.393.4 latest @lukehagar/discoursejs",
 } as const;